Added images and basic formating
The basic page layout is approximately correct. A good jumping off point
at least. Code should be reviewed to make sure it's not too verbose.

Found some images to use in the site. Not all of them are suitable for a
public site, but this is mostly just a project to test my skills, so I'm
fine with it for now. Especially considering that I don't plan on using
this seriously